chore: satisfy the debt ratchet for the component-parts change
research_workspace sheds the dead post-result surface replay (new turns
never produce a2ui_surfaces) to sit at its 679-line ceiling, and the
component-part turn tests move to their own module so test_turn_engine
stays under the 500-line new-file limit. The legacy replay test now
expects one fewer duplicate delta.
